Initial commit for SadCL from Leland Batey.
This is the initial commit for SadCL. So far, there's a skeleton of TargetManager as well as the beginnings of unit-testing.

Lot's of work still remains, but and there are a lot of junk files in this (temporary stuff I think) that is making it's way in. Oh well, I'll just have to clean that out later.