Added ability to load target files
I've added the ability on the backend to load target files, as well as the GUI for loading target files.
As part of this, I had to make a small change to SadCL, in the TargetManager, so that it could expose the master list of all targets.

Currently, the GUI is atrocious but functional. If a target file cannot be loaded (incorrect syntax/incorrect path) a simple (and somewhat unhelpful) error will be shown. On successfully loading a file, a simple success message is displayed.