Added buttons to aim and kill targets
The buttons allow you to aim or kill targets that are loaded. Right now, that's specified by the number that is the index of the target in the list of targets. There's no bounds checking, so that could be bad. I'll do that later.

I think the last thing I'd need to do is to be able to see the status of the turret, but that's a problem for another day...