Fixed duplicate namespacing
Somehow we'd accidentally created two different namespaces for our ViewModels. There was `SadCLGUI.GUI_ViewModels` and `SadCLGUI.ViewModels` . It's now been fixed so that theres only one namespace, `SadCLGUI.ViewModels` .

`ViewModels` was chosen because it's the shorter of the two, and it seemed somewhat redundant to have `GUI` in the viewmodel namespace when the entire solution is for the gui.